backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pytest
httpx== 0.27.2
sqlalchemy
passlib[bcrypt]
python-multipart
python-jose[cryptography]
psycopg2-binary